chore: move uploadoptions to schemas
